requirements.txt
tensorflow==2.14.0
streamlit==1.28.2
numpy==1.24.3
pillow==9.5.0